+++ screens/Overview_Edit.js
@@ -17,6 +17,10 @@ import DateTimePicker from "@react-native-community/datetimepicker";
 import { Ionicons } from "@expo/vector-icons";
 import { updateFinanceItem } from "../js/redux/actions/Finance";
 import firebase from "../js/Firebase";
+import {
+  selectablePaymentMethods,
+  uploadImage,
+} from "../js/Helper";
 import NewEntry_Camera from "../components/NewEntry_Camera";
 import MoneyInput from "../components/MoneyInput";
 import { ObjectItemPicker, StringItemPicker } from "../components/ItemPicker";
@@ -318,13 +322,6 @@ function Overview_Edit(props) {
   );
 }
 
-const selectablePaymentMethods = [
-  { label: "Barzahlung", value: "cash" },
-  { label: "EC-Karte", value: "debit-card" },
-  { label: "Kreditkarte", value: "credit-card" },
-  { label: "PayPal", value: "paypal" },
-];
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -368,41 +365,6 @@ const styles = StyleSheet.create({
     paddingVertical: 5,
   },
 });
-
-async function uploadImage(uri) {
-  const response = await fetch(uri);
-  const imageBlob = await response.blob();
-
-  const metadata = {
-    contentType: "image/jpeg",
-  };
-  const name = generateUUID() + "-media.jpg";
-  const imageRef = firebase.storage.ref().child("images/" + name);
-
-  const snapshot = await imageRef.put(imageBlob, metadata);
-  const downloadURL = await snapshot.ref.getDownloadURL();
-
-  return downloadURL;
-}
-
-function generateUUID() {
-  // Public Domain/MIT
-  var d = new Date().getTime(); //Timestamp
-  var d2 = (performance && performance.now && performance.now() * 1000) || 0; //Time in microseconds since page-load or 0 if unsupported
-  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, function (c) {
-    var r = Math.random() * 16; //random number between 0 and 16
-    if (d > 0) {
-      //Use timestamp until depleted
-      r = (d + r) % 16 | 0;
-      d = Math.floor(d / 16);
-    } else {
-      //Use microseconds since page-load if supported
-      r = (d2 + r) % 16 | 0;
-      d2 = Math.floor(d2 / 16);
-    }
-    return (c === "x" ? r : (r & 0x3) | 0x8).toString(16);
-  });
-}
 
 const mapStateToProps = (state) => {
   return { currentUser: state.user };
